Move the clearInterval call top level in the elapsed duration effect to solve the frustrating duration UI flashing on seek
