+++ features/players/components/twitch-player/TwitchPlayerControls.tsx
@@ -17,6 +17,7 @@ import PauseIcon from "icons/PauseIcon";
 import SettingsGearIcon from "icons/SettingsGearIcon";
 import { TwitchPlayerSettingsMenu } from "./TwitchPlayerSettingsMenu";
 import * as React from "react";
+import { useVideoTime } from "features/players/hooks/useVideoTimer";
 
 interface TwitchPlayerControlsProps {
   player: Twitch.Player;
@@ -43,31 +44,11 @@ export const TwitchPlayerControls = ({
   playerMuted,
   projectedTime,
 }: TwitchPlayerControlsProps) => {
-  const durationInterval = React.useRef<null | NodeJS.Timer>(null);
-
-  // A constantly updated duration state to provide a video duration elapsed to the UI
-  const [elapsedDuration, setElapsedDuration] = useState("");
-
   // Controls display of video quality settings menu
   const [showSettingsMenu, setShowSettingsMenu] = useState(false);
+  const durationInterval = React.useRef<null | NodeJS.Timer>(null);
 
-  // An initial render effect only to avoid a 1 second delay in showing elapsed time
-  useEffect(() => {
-    const elapsedTime = player.getCurrentTime();
-    setElapsedDuration(formatElapsedTime(elapsedTime));
-  }, [player]);
-
-  useEffect(() => {
-    if (projectedTime) {
-      clearInterval(durationInterval.current as NodeJS.Timer);
-      setElapsedDuration(formatElapsedTime(projectedTime));
-    } else {
-      durationInterval.current = setInterval(() => {
-        const elapsedTime = player.getCurrentTime();
-        setElapsedDuration(formatElapsedTime(elapsedTime));
-      }, 1000);
-    }
-  }, [player, projectedTime]);
+  const { elapsedDuration } = useVideoTime(player, projectedTime);
 
   // Use this function in any position where the user's focus should return to the video
   const releaseFocus = () => {
